[some cleaning]

+++ src/main.cpp
@@ -165,6 +165,9 @@ int main(int argc, char* argv[])
            // output_vd(vd,place);    //output particle properties
             
             updateParticleProperties(vd, place, dt, H, turb, simulation);
+
+            // smooth P
+
             double cfl_temp = (dt/H) * (vd.template getProp<i_velocity>(p)[0]+vd.template getProp<i_velocity>(p)[1]+vd.template getProp<i_velocity>(p)[2]);
             cfl = std::max(abs(cfl_temp),cfl);       //look for max not average. needs to be absolute.  should hold
        
@@ -188,7 +191,7 @@ int main(int argc, char* argv[])
         NN = vd.getCellList(r_cut);
 
         // collect statistics on the configuration
-        //if ((i+1) % simulation.frame == 0)
+        if ((i+1) % simulation.frame == 0)
         {
             // Write the particle position for visualization (Without ghost)
             vd.deleteGhost();
